forel_start

+++ sotrudnik.cpp
@@ -1,6 +1,6 @@
 #include "sotrudnik.h"
 
-Sotrudnik::Sotrudnik(vector<int> abilities, int norm, int busy, bool entry)
+Sotrudnik::Sotrudnik(vector<double> abilities, int norm, int busy, bool entry)
 {
     this->abilities = abilities;
     this->norm = norm;
